requirements.txt
duckduckgo-search>=5.3.0
google-genai>=0.2.0
schedule>=1.2.1
beautifulsoup4>=4.12.0
requests>=2.31.0
